Files modified: 1
Modified 1 file:src/skin/graphic/decorator/RoundRectGraphic.swift

+++ src/skin/graphic/decorator/RoundRectGraphic.swift
@@ -13,6 +13,6 @@ class RoundRectGraphic:GraphicDecoratable{//adds round-rectangular path
         let h:CGFloat = getGraphic().height
         //Swift.print("h: " + "\(h)")
         //Swift.print("fillet.topLeft: " + "\(fillet.topLeft)")
-        getGraphic().path = CGPathParser.roundRect(0,0,CGFloat(w), CGFloat(h),CGFloat(fillet.topLeft), CGFloat(fillet.topRight), CGFloat(fillet.bottomLeft), CGFloat(fillet.bottomRight))//Shapes
+        getGraphic().path = CGPathParser.roundRect(0,0,w, h,CGFloat(fillet.topLeft), CGFloat(fillet.topRight), CGFloat(fillet.bottomLeft), CGFloat(fillet.bottomRight))//Shapes
     }
 }
